% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProvider.r
\name{DataProvider}
\alias{DataProvider}
\title{DataProvider constructor}
\usage{
DataProvider(parent = NULL)
}
\arguments{
\item{parent}{a back-up DataProvider. If the data is not found in the current provider, the back-up provider is searched.}
}
\value{
an object of type DataProvider
}
\description{
Constructor
}
