% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fInstrument-package.r
\docType{package}
\name{fInstrument-package}
\alias{fInstrument-package}
\title{fInstrument Package}
\description{
The fInstrument package is useful for performing calculations on financial instruments defined in Rmetrics.
}
\author{
P. Henaff
}
