% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProvider.r
\name{getData}
\alias{getData}
\alias{getData,DataProvider-method}
\title{getData method}
\usage{
getData(object, instrument, phenomenon, dtObs)

\S4method{getData}{DataProvider}(object, instrument, phenomenon, dtObs)
}
\arguments{
\item{instrument}{instrument name}

\item{phenomenon}{phenomenon name}

\item{dtObs}{observation date}
}
\value{
a vector of value(s)
}
\description{
Extract data from a DataProvider object
}
