% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fInstrument.r
\name{getValue}
\alias{getValue}
\alias{getValue,fInstrument-method}
\title{GetValue}
\usage{
getValue(object, selection, dtCalc, env = NULL)

\S4method{getValue}{fInstrument}(object, selection, dtCalc, env = NULL)
}
\arguments{
\item{selection}{The type of calculation ('Value', 'Delta', 'Gamma', 'Vega')}

\item{dtCalc}{the calculation date}

\item{env}{the data provider}
}
\value{
a time series of results
}
\description{
Invoque a calculation on a financial instrument
}
