\name{setData}
\alias{setData}
\title{insert data into DataProvider objectsetData}
\arguments{
  \item{instrument}{(char) instrument name}

  \item{phenomenon}{(char) phenomenon name}

  \item{dtObs}{(date) observation date}

  \item{value}{(numeric) value}
}
\description{
  Insert data in a DataProvider object
}

