% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fInstrument.r
\name{show,fInstrument-method}
\alias{show,fInstrument-method}
\title{Display a description of the instrument}
\usage{
\S4method{show}{fInstrument}(object)
}
\value{
A description of the instrument
}
\description{
Show
}
\details{
Provides a description of the instruments (i.e. the value of the main parameters
found in the constructor
}
