\name{getValue}
\alias{getValue}
\title{GetValue}
\arguments{
  \item{selection}{The type of calculation ('Value',
  'Delta', 'Gamma', 'Vega')}

  \item{dtCalc}{the calculation date}

  \item{env}{the data provider}
}
\value{
  a time series of results
}
\description{
  Invoque a calculation on a financial instrument
}

