% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProvider.r
\name{setData}
\alias{setData}
\alias{setData,DataProvider-method}
\title{insert data into DataProvider objectsetData}
\usage{
setData(object, instrument, phenomenon, dtObs, value)

\S4method{setData}{DataProvider}(object, instrument, phenomenon, dtObs, value)
}
\arguments{
\item{instrument}{(char) instrument name}

\item{phenomenon}{(char) phenomenon name}

\item{dtObs}{(date) observation date}

\item{value}{(numeric) value}
}
\description{
Insert data in a DataProvider object
}
